pyproject.toml
[tool.poetry]
name = "kafka-structure-streaming"
version = "0.1.0"
description = ""
authors = ["Adeola"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
confluent-kafka = "^2.13.0"
python-dotenv = "^1.2.1"
pyspark = "3.5.1"
notebook = "^7.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
